+++ TiendaDeClases/app/src/main/java/desarrollomobile/tiendadeclases/tiendadeclases/Activities/MainActivity.kt
@@ -1,4 +1,4 @@
-package desarrollomobile.tiendadeclases.tiendadeclases
+package desarrollomobile.tiendadeclases.tiendadeclases.Activities
 
 import android.net.Uri
 import android.os.Bundle
@@ -6,6 +6,10 @@ import android.os.Handler
 import android.support.v4.app.Fragment
 import android.support.v4.app.FragmentTransaction
 import android.support.v7.app.AppCompatActivity
+import desarrollomobile.tiendadeclases.tiendadeclases.Fragments.Login.LoginFragment
+import desarrollomobile.tiendadeclases.tiendadeclases.R
+import desarrollomobile.tiendadeclases.tiendadeclases.Fragments.Login.SplashFragment
+import desarrollomobile.tiendadeclases.tiendadeclases.Fragments.Login.Tutorial1Fragment
 
 
 class MainActivity : AppCompatActivity(), SplashFragment.OnFragmentInteractionListener, Tutorial1Fragment.OnFragmentInteractionListener, LoginFragment.OnFragmentInteractionListener {
