+++ TiendaDeClases/app/src/main/java/desarrollomobile/tiendadeclases/tiendadeclases/Fragments/Login/Tutorial1Fragment.kt
@@ -1,4 +1,4 @@
-package desarrollomobile.tiendadeclases.tiendadeclases
+package desarrollomobile.tiendadeclases.tiendadeclases.Fragments.Login
 
 import android.content.Context
 import android.net.Uri
@@ -9,6 +9,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
+import desarrollomobile.tiendadeclases.tiendadeclases.R
 
 
 // TODO: Rename parameter arguments, choose names that match
