+++ TiendaDeClases/app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity">
+    tools:context=".Activities.HomeActivity">
 
     <FrameLayout
         android:id="@+id/HomeFrame"
