+++ TiendaDeClases/app/src/main/res/layout/fragment_categories.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CategoriesFragment">
+    tools:context=".Fragments.Home.CategoriesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
