+++ TiendaDeClases/app/src/main/res/layout/fragment_login.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".LoginFragment">
+    tools:context=".Fragments.Login.LoginFragment">
 
     <!-- TODO: Update blank fragment layout -->
