+++ TiendaDeClases/app/src/main/res/layout/fragment_messages.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MessagesFragment">
+    tools:context=".Fragments.Home.MessagesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
