+++ TiendaDeClases/app/src/main/res/layout/fragment_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".SplashFragment">
+    tools:context=".Fragments.Login.SplashFragment">
 
     <!-- TODO: Update blank fragment layout -->
