+++ TiendaDeClases/app/src/main/res/layout/fragment_tutorial1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Tutorial1Fragment">
+    tools:context=".Fragments.Login.Tutorial1Fragment">
 
     <!-- TODO: Update blank fragment layout -->
